requirements.txt
pillow
pandas
pydicom
nibabel
tqdm
torchvision>=0.2.1
dominate>=2.3.1
visdom>=0.1.8.
simpleITK==1.2.4
torchsummary
torchsummaryX
scikit-image
scikit-learn